refactor(assignments): move hooks e tipos do LotteryDialog para arquivos próprios
useLotteryStats, useLotteryParams (com o tipo LotteryParamsState) e os
tipos LotteryMember/LotteryStats/CodingsFilterMode saem do LotteryDialog.tsx
para módulos co-localizados (padrão compare/), sem mudança de lógica.
Primeiro passo da decomposição do no-giant-component (#357).
