+++ frontend/src/components/assignments/useLotteryStats.ts
@@ -1,0 +1,30 @@
+import { useEffect, useState } from "react";
+import { getLotteryDocStats } from "@/actions/assignments";
+import type { LotteryStats } from "./lottery-dialog-types";
+
+// Stats de elegibilidade, recarregadas a cada abertura do dialog —
+// um sorteio muda atribuições/lotes, então reabrir com stats da
+// abertura anterior mentiria na contagem de elegíveis. Os dois campos
+// (dados + erro) vivem num único objeto de estado para que o effect
+// faça uma única chamada de setter por branch (evita cascading set-state).
+export function useLotteryStats(projectId: string, open: boolean) {
+  const [statsState, setStatsState] = useState<{
+    data: LotteryStats | null;
+    error: boolean;
+  }>({ data: null, error: false });
+  useEffect(() => {
+    if (!open) return;
+    let cancelled = false;
+    getLotteryDocStats(projectId)
+      .then((s) => {
+        if (!cancelled) setStatsState({ data: s, error: false });
+      })
+      .catch(() => {
+        if (!cancelled) setStatsState((prev) => ({ ...prev, error: true }));
+      });
+    return () => {
+      cancelled = true;
+    };
+  }, [open, projectId]);
+  return { stats: statsState.data, statsError: statsState.error };
+}
